Round the floats

+++ app/graphql/types/analytics/page.rb
@@ -8,7 +8,7 @@ module Types
       field :path, String, null: false
       field :count, Integer, null: false
       field :avg, Integer, null: false
-      field :percentage, Integer, null: false
+      field :percentage, Float, null: false
     end
   end
 end
